requirements.txt
langchain
langgraph
streamlit
python-dotenv
supabase
openai
requests
